+++ integration_test/cases/browser/invalid_selectors_test.exs
@@ -15,5 +15,11 @@ defmodule Wallaby.Integration.Browser.InvalidSelectorsTest do
         assert_has(session, css("checkbox:foo"))
       end
     end
+
+    test "refute_has returns an exception", %{session: session} do
+      assert_raise Wallaby.QueryError, ~r/The css 'checkbox:foo' is not a valid query/, fn ->
+        refute_has(session, css("checkbox:foo"))
+      end
+    end
   end
 end